Adding error message to `ABCToolBoxRejectWorker`.
If the prior file is too large, ABCestimator will seg fault but still
exit with code 0. Thus, the code crashes when an instance of
`ABCToolBoxRejectWorker` goes to post process the posterior file.
This commit adds a check for the posterior files existance during
the `_post_processing` method, and raises an Exception if it does
not exist, and prints the stderr from the ABCestimator subprocess.

This should create clearer error messaging when ABCestimator chokes
on a prior file that is too large.

+++ pymsbayes/workers.py
@@ -789,6 +789,10 @@ class ABCToolBoxRejectWorker(Worker):
 
     def _post_process(self):
         post_path = self.output_prefix + 'BestSimsParamStats_Obs0.txt'
+        if not os.path.exists(post_path):
+            raise Exception('{0} did not produce a posterior file.\n'
+                    'Here is the stderr from the subprocess:\n{1}'.format(
+                            self.name, self.get_stderr()))
         # remove extra two columns added by abctoolbox
         with open(self.posterior_path, 'w') as o:
             with open(post_path, 'rU') as i:
